requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
pydantic-settings==2.5.2

langchain==0.3.1
langchain-community==0.3.1
langchain-openai==0.2.1
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
chromadb==0.5.5
sentence-transformers==3.1.1
pypdf==5.0.1
docx2txt==0.8

python-telegram-bot==21.6
